Don't create interfaces from entity types without bundles.
